@@ -1,32 +1,31 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "npm run development -- --watch",
-        "watch-poll": "npm run watch -- --watch-poll",
-        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "dev": "vite",
+        "build": "vite build"
     },
     "devDependencies": {
         "@babel/preset-react": "^7.0.0",
-        "axios": "^0.21.2",
-        "bootstrap": "^4.0.0",
+        "@vitejs/plugin-react": "^2.1.0",
+        "@vitejs/plugin-vue": "^3.1.0",
+        "axios": "^0.27.2",
+        "bootstrap": "^4.6.0",
         "cross-env": "^7.0",
         "jquery": "^3.2",
-        "laravel-mix": "^5.0.1",
+        "laravel-vite-plugin": "^0.6.1",
         "lodash": "^4.17.13",
         "popper.js": "^1.12",
-        "react": "^16.2.0",
-        "react-dom": "^16.2.0",
-        "resolve-url-loader": "^3.1.0",
-        "sass": "^1.15.2",
-        "sass-loader": "^8.0.0"
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "resolve-url-loader": "^5.0.0",
+        "sass": "^1.55.0",
+        "sass-loader": "^13.1.0",
+        "vite": "^3.1.4"
     },
     "dependencies": {
-        "@fortawesome/fontawesome-free": "^5.13.0",
-        "admin-lte": "^3.0.4",
-        "sweetalert2": "^9.10.12"
+        "@fortawesome/fontawesome-free": "^6.2.0",
+        "admin-lte": "^3.2.0",
+        "dateformat": "^5.0.3",
+        "sweetalert2": "^11.4.37"
     }
 }
