@@ -1,4 +1,5 @@
-window._ = require("lodash");
+import _ from 'lodash';
+window._ = _;
 
 /**
  * We'll load jQuery and the Bootstrap jQuery plugin which provides support
@@ -6,15 +7,28 @@ window._ = require("lodash");
  * code may be modified to fit the specific needs of your application.
  */
 
-window.Swal = require("sweetalert2");
+import Swal from 'sweetalert2';
+window.Swal = Swal;
 
-try {
-    window.Popper = require("popper.js").default;
-    window.$ = window.jQuery = require("jquery");
+import Popper from 'popper.js';
+window.Popper = Popper;
 
-    require("bootstrap");
-    require("admin-lte");
-} catch (e) {}
+import jQuery from 'jquery';
+window.$ = window.jQuery = jQuery;
+
+import DataTable from 'datatables.net-bs4';
+DataTable(window, window.$);
+
+import 'bootstrap';
+import 'admin-lte';
+
+window.SwalWithBootstrap = Swal.mixin({
+  customClass: {
+    confirmButton: 'btn btn-primary mx-3',
+    cancelButton: 'btn btn-danger mx-3',
+  },
+  buttonsStyling: false,
+});
 
 /**
  * We'll load the axios HTTP library which allows us to easily issue requests
@@ -22,7 +36,8 @@ try {
  * CSRF token as a header based on the value of the "XSRF" token cookie.
  */
 
-window.axios = require("axios");
+import axios from 'axios';
+window.axios = axios;
 
 window.axios.defaults.headers.common["X-Requested-With"] = "XMLHttpRequest";
 
