@@ -10,9 +10,11 @@
     <!-- Ionicons -->
     <link rel="stylesheet" href="https://code.ionicframework.com/ionicons/2.0.1/css/ionicons.min.css">
     <!-- overlayScrollbars -->
-    <link rel="stylesheet" href="{{ asset('css/app.css') }}">
+    <!-- <link rel="stylesheet" href="{{ asset('css/app.css') }}"> -->
     <!-- Google Font: Source Sans Pro -->
     <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700" rel="stylesheet">
+
+    @vite(['resources/sass/app.scss', 'resources/js/app.js'])
     @yield('css')
     <script>
         window.APP = <?php echo json_encode([
@@ -63,8 +65,7 @@
         <!-- /.control-sidebar -->
     </div>
     <!-- ./wrapper -->
-
-    <script src="{{ asset('js/app.js') }}"></script>
+    <!-- <script src="{{ asset('js/app.js') }}"></script> -->
     @yield('js')
 </body>
 
